refactor: consolidate buildExplainResult into internal/rules package
Consolidate the duplicated `buildExplainResult` function and its
associated types (`ExplainResult`, `ExplainField`, `ExplainError`)
into the `internal/rules` package as `NewExplainResult` and
exported types.

This improves maintainability and ensures that both the CLI and
the MCP tool use the same logic for tracing field origins and
validation errors.

- Created `internal/rules/explain.go` with shared logic and types.
- Updated `cmd/rootline/explain.go` to use `rules.NewExplainResult`.
- Updated `internal/mcp/tools.go` to use `rules.NewExplainResult`.
- Moved and updated unit tests to `internal/rules/explain_test.go`.
