+++ cmd/rootline/explain.go
@@ -25,33 +25,6 @@ func init() {
 	rootCmd.AddCommand(explainCmd)
 }
 
-// ExplainResult is the versioned JSON output for the explain command.
-type ExplainResult struct {
-	Version   int            `json:"version"`
-	Kind      string         `json:"kind"`
-	Path      string         `json:"path"`
-	StemChain []string       `json:"stem_chain"`
-	Fields    []ExplainField `json:"fields"`
-	Errors    []ExplainError `json:"errors,omitempty"`
-}
-
-// ExplainField traces a single field's value and origin.
-type ExplainField struct {
-	Name       string `json:"name"`
-	Value      any    `json:"value"`
-	Origin     string `json:"origin"`
-	Source     string `json:"source,omitempty"`
-	Expression string `json:"expression,omitempty"`
-}
-
-// ExplainError traces a validation error to its source rule.
-type ExplainError struct {
-	Rule     string `json:"rule"`
-	Field    string `json:"field"`
-	Message  string `json:"message"`
-	Source   string `json:"source"`
-	Severity string `json:"severity"`
-}
 
 func runExplain(cmd *cobra.Command, args []string) error {
 	ctx := cmd.Context()
@@ -122,7 +95,7 @@ func runExplain(cmd *cobra.Command, args []string) error {
 	}
 
 	// Build explain result.
-	result := buildExplainResult(file, entries, effective, record, valErrs)
+	result := rules.NewExplainResult(file, entries, effective, record, valErrs)
 
 	if outputFormat == "table" {
 		return renderExplainTable(cmd, result)
@@ -130,102 +103,7 @@ func runExplain(cmd *cobra.Command, args []string) error {
 	return outputJSON(cmd, result, false)
 }
 
-func buildExplainResult(
-	path string,
-	entries []rules.StemEntry,
-	effective *rules.StemFile,
-	record *extract.Record,
-	valErrs []rules.ValidationError,
-) *ExplainResult {
-	chain := make([]string, len(entries))
-	for i, e := range entries {
-		chain[i] = e.Path
-	}
-
-	var fields []ExplainField
-
-	// Frontmatter fields.
-	fmKeys := sortedMapKeys(record.Frontmatter)
-	for _, k := range fmKeys {
-		f := ExplainField{
-			Name:   k,
-			Value:  record.Frontmatter[k],
-			Origin: "frontmatter",
-		}
-		// Check if this field is defined in schema.
-		if effective != nil {
-			if sf, ok := effective.Schema[k]; ok {
-				f.Source = sf.Source
-			}
-		}
-		fields = append(fields, f)
-	}
-
-	// Schema fields not present in frontmatter (missing required, defaults).
-	if effective != nil {
-		for name, sf := range effective.Schema {
-			if _, exists := record.Frontmatter[name]; exists {
-				continue
-			}
-			f := ExplainField{
-				Name:   name,
-				Value:  nil,
-				Origin: "schema",
-				Source: sf.Source,
-			}
-			if sf.Default != "" {
-				f.Value = sf.Default
-			}
-			fields = append(fields, f)
-		}
-	}
-
-	// Derived fields (from derive and aggregate).
-	if effective != nil {
-		derivedKeys := sortedMapKeys(record.Derived)
-		for _, k := range derivedKeys {
-			f := ExplainField{
-				Name:   k,
-				Value:  record.Derived[k],
-				Origin: "derived",
-			}
-			if exprVal, ok := effective.Derive[k]; ok {
-				if exprStr, ok := exprVal.(string); ok {
-					f.Expression = exprStr
-				}
-			} else if exprVal, ok := effective.Aggregate[k]; ok {
-				f.Origin = "aggregate"
-				if exprStr, ok := exprVal.(string); ok {
-					f.Expression = exprStr
-				}
-			}
-			fields = append(fields, f)
-		}
-	}
-
-	// Validation errors.
-	var explainErrs []ExplainError
-	for _, ve := range valErrs {
-		explainErrs = append(explainErrs, ExplainError{
-			Rule:     ve.Rule,
-			Field:    ve.Field,
-			Message:  ve.Message,
-			Source:   ve.Source,
-			Severity: ve.Severity,
-		})
-	}
-
-	return &ExplainResult{
-		Version:   1,
-		Kind:      "rootline/explain",
-		Path:      path,
-		StemChain: chain,
-		Fields:    fields,
-		Errors:    explainErrs,
-	}
-}
-
-func renderExplainTable(cmd *cobra.Command, r *ExplainResult) error {
+func renderExplainTable(cmd *cobra.Command, r *rules.ExplainResult) error {
 	_, _ = fmt.Fprintf(cmd.OutOrStdout(), "File: %s\n", r.Path)
 	_, _ = fmt.Fprintf(cmd.OutOrStdout(), "Stem chain: %v\n\n", r.StemChain)
